it_opss/requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.24.0
requests>=2.28.0
python-dotenv>=1.0.0
openai>=1.3.0
httpx>=0.24.0
matplotlib>=3.7.0
